In the middle if Implementing some new strategies for the YatzyAgent.
