--- src/enjug/erijan/games/yatzy/ButtonsPanel.java
@@ -1,10 +1,0 @@
-package enjug.erijan.games.yatzy;
-
-import javax.swing.*;
-
-/**
- * Created by Jan Eriksson on 19/11/15.
- */
-public class ButtonsPanel extends JPanel {
-
-}
