+++ src/enjug/erijan/games/yatzy/RollingControlling.java
@@ -3,7 +3,7 @@ package enjug.erijan.games.yatzy;
 /**
  * Created by Jan Eriksson on 16/11/15.
  */
-public interface RollingBehaviour {
-  String rollActiveDice();
+public interface RollingControlling {
+  void rollActiveDice();
   int rollsLeft();
 }
