+++ src/enjug/erijan/games/yatzy/StateInfo.java
@@ -1,0 +1,9 @@
+package enjug.erijan.games.yatzy;
+
+/**
+ * Created by Jan Eriksson on 20/11/15.
+ */
+public interface StateInfo {
+  String getStateMessage();
+  void setStateMessage(String stateMessage);
+}
