+++ src/enjug/erijan/games/yatzy/YatzyRolling.java
@@ -1,16 +1,33 @@
 package enjug.erijan.games.yatzy;
 
+import enjug.erijan.games.util.DiceHandler;
+
+import java.util.List;
+
 /**
  * Created by Jan Eriksson on 16/11/15.
  */
-public class YatzyRolling implements RollingBehaviour {
+public class YatzyRolling implements RollingControlling {
+
+  private static final int maxRerolls = 2;
+
+  private DiceHandler diceHandler;
+  private List<Player> players;
+  private int reRolls;
+//  Map<String,Integer> rolls
+
+  YatzyRolling(List<Player> players, DiceHandler diceHandler) {
+    this.players = players;
+    this.diceHandler = diceHandler;
+  }
+
   @Override
-  public String rollActiveDice() {
-    return null;
+  public void rollActiveDice() {
+    diceHandler.rollActiveDice();
   }
 
   @Override
   public int rollsLeft() {
-    return 0;
+    return maxRerolls - reRolls;
   }
 }
